// FrontEnd/src/app/models/academic.ts
export class Academics {
    id: number = 0;
    usuarios_id: number = 0;
    institucion: string = "";
    titulo: string = "";
    locacion: string = "";
    habilidades: string[] = []; //agregar a la base de datos ¿sí o no?
    desde: string = "";
    hasta: string = "";
    logo: string = ""
    logo_url: string = "";
    logo_public_id: string = ""; 
    usuario_id: number = 0;
}