// FrontEnd/src/app/models/experience.ts
export class Experience {
    id: number = 0;
    cargo: string = "";
    tarea: string[] = []; //en el get hacer el split y guardarlo en las tareas
    desde: string = "";
    hasta: string = "";
    logo: string = "";
    logo_url: string = "";
    logo_public_id: string = ""; 
    empresa: string = "";
    reftelef: string = "";
    refnombre: string = "";
    usuarios_id: number = 0;
}