// FrontEnd/src/app/models/projects.ts
export class Project {
    id: number = 0;
    usuarios_id: number = 0;
    nombre: string = "";
    descripcion: string = "";
    desde: string = "";
    hasta: string = "";
    enlace: string = "";
    sitio: string = "";
    logo: string = "";
    logo_url: string = "";
    logo_public_id: string = ""; 
}